вынес абстракции источников данных
