+++ lib/feature/data/datasources/contacts_datasource/contacts_local_datasource.dart
@@ -1,20 +1,9 @@
 import 'dart:developer';
 
+import 'package:cportal_flutter/feature/data/i_datasource/i_local_datasource/i_contacts_local_datasource.dart';
 import 'package:cportal_flutter/feature/data/models/contacts_model.dart';
 import 'package:flutter/foundation.dart';
 import 'package:hive/hive.dart';
-
-abstract class IContactsLocalDataSource {
-  /// Извлекаем [ContactsModel] из кеша
-  ///
-  /// Пробрасываем все ошибки через [CacheException]
-  Future<ContactsModel> fetchContactsFromCache();
-
-  /// Сохраняем [ContactsModel] в кэш
-  ///
-  /// Пробрасываем все ошибки через [CacheException]
-  Future<void> contactsToCache(ContactsModel contacts);
-}
 
 class ContactsLocalDataSource implements IContactsLocalDataSource {
   final HiveInterface hive;
