+++ lib/feature/data/datasources/contacts_datasource/contacts_remote_datasource.dart
@@ -1,25 +1,14 @@
 import 'dart:convert';
 import 'dart:developer';
 import 'package:cportal_flutter/app_config.dart';
+import 'package:cportal_flutter/feature/data/i_datasource/i_local_datasource/i_contacts_local_datasource.dart';
+import 'package:cportal_flutter/feature/data/i_datasource/i_remote_datasource/i_contacts_remote_datasource.dart';
 import 'package:cportal_flutter/feature/domain/entities/profile_entity.dart';
 import 'package:dio/dio.dart';
 
 import 'package:cportal_flutter/core/error/server_exception.dart';
 import 'package:cportal_flutter/core/error/failure.dart';
-import 'package:cportal_flutter/feature/data/datasources/contacts_datasource/contacts_local_datasource.dart';
 import 'package:cportal_flutter/feature/data/models/contacts_model.dart';
-
-abstract class IContactsRemoteDataSource {
-  /// Обращается к эндпойнту .....
-  ///
-  /// Пробрасываем ошибки через [ServerException]
-  Future<ContactsModel> fetchContacts(int page);
-
-  /// Обращается к эндпойнту .....
-  ///
-  /// Пробрасываем ошибки через [ServerException]
-  Future<List<ProfileEntity>> fetchContactsBySearch(String query);
-}
 
 class ContactsRemoteDataSource implements IContactsRemoteDataSource {
   final IContactsLocalDataSource localDataSource;
@@ -32,8 +21,7 @@ class ContactsRemoteDataSource implements IContactsRemoteDataSource {
 
   @override
   Future<ContactsModel> fetchContacts(int page) async {
-    final String baseUrl =
-        '${AppConfig.apiUri}/cportal/hs/api/contacts/1.0/?page=$page';
+    final String baseUrl = '${AppConfig.apiUri}/cportal/hs/api/contacts/1.0/?page=$page';
     try {
       final response = await dio.get<String>(baseUrl);
 
@@ -52,8 +40,7 @@ class ContactsRemoteDataSource implements IContactsRemoteDataSource {
 
   @override
   Future<List<ProfileEntity>> fetchContactsBySearch(String query) async {
-    final String baseUrl =
-        '${AppConfig.apiUri}/cportal/hs/api/contacts/1.0/?q=$query';
+    final String baseUrl = '${AppConfig.apiUri}/cportal/hs/api/contacts/1.0/?q=$query';
     try {
       final response = await dio.get<String>(baseUrl);
