+++ lib/feature/data/datasources/user_datasource/user_remote_datasource.dart
@@ -1,14 +1,8 @@
 import 'package:cportal_flutter/core/error/server_exception.dart';
 import 'package:cportal_flutter/core/error/failure.dart';
-import 'package:cportal_flutter/feature/data/datasources/user_datasource/user_local_datasource.dart';
+import 'package:cportal_flutter/feature/data/i_datasource/i_local_datasource/i_user_local_datasource.dart';
+import 'package:cportal_flutter/feature/data/i_datasource/i_remote_datasource/i_user_remote_datasource.dart';
 import 'package:cportal_flutter/feature/data/models/user_model.dart';
-
-abstract class IUserRemoteDataSource {
-  /// Обращается к эндпойнту .....
-  ///
-  /// Пробрасываем все ошибки через [ServerException]
-  Future<UserModel> login(String connectongCode);
-}
 
 class UserRemoteDataSource implements IUserRemoteDataSource {
   final IUserLocalDataSource localDatasource;
