+++ lib/feature/data/i_datasource/i_local_datasource/i_contacts_local_datasource.dart
@@ -1,0 +1,13 @@
+import 'package:cportal_flutter/feature/data/models/contacts_model.dart';
+
+abstract class IContactsLocalDataSource {
+  /// Извлекаем [ContactsModel] из кеша
+  ///
+  /// Пробрасываем все ошибки через [CacheException]
+  Future<ContactsModel> fetchContactsFromCache();
+
+  /// Сохраняем [ContactsModel] в кэш
+  ///
+  /// Пробрасываем все ошибки через [CacheException]
+  Future<void> contactsToCache(ContactsModel contacts);
+}
