+++ lib/feature/data/i_datasource/i_local_datasource/i_pin_code_local_datasource.dart
@@ -1,0 +1,11 @@
+abstract class IPinCodeLocalDataSource {
+  /// Записывает [String] в кэш
+  ///
+  /// Пробрасываем все ошибки через [CacheException]
+  Future<String> writePin(String pinCode);
+
+  /// Запрашивает есть ли ПИН в кэше
+  ///
+  /// Пробрасываем все ошибки через [CacheException]
+  Future<String?> getPin();
+}
