+++ lib/feature/data/i_datasource/i_remote_datasource/i_filter_remote_datasource.dart
@@ -1,0 +1,9 @@
+import 'package:cportal_flutter/feature/data/models/filter_model.dart';
+
+abstract class IFilterRemoteDataSource {
+  // Пробрасываем ошибки через [ServerException].
+  Future<FilterResponseModel> fetchContactsFilters();
+
+  // Пробрасываем ошибки через [ServerException].
+  Future<FilterResponseModel> fetchDeclarationsFilters();
+}
