+++ lib/feature/data/i_datasource/i_remote_datasource/i_news_remote_datasource.dart
@@ -1,0 +1,23 @@
+import 'package:cportal_flutter/feature/data/models/news_model.dart';
+
+abstract class INewsRemoteDataSource {
+  /// Обращается к эндпойнту .....
+  /// Возвращает [NewsModel]
+  /// Пробрасываем ошибки через [ServerException]
+  Future<NewsModel> fetchNews(int page);
+
+  /// Обращается к эндпойнту .....
+  /// Возвращает [NewsModel]
+  /// Пробрасываем ошибки через [ServerException]
+  Future<NewsModel> fetchNewsByCategory(int page, String category);
+
+  /// Обращается к эндпойнту .....
+  /// Возвращает [NewsModel]
+  /// Пробрасываем ошибки через [ServerException]
+  Future<NewsModel> fetchQuastions(int page);
+
+  /// Обращается к эндпойнту .....
+  /// Возвращает [NewsModel]
+  /// Пробрасываем ошибки через [ServerException]
+  Future<NewsModel> fetchQuastionsByCategory(int page, String category);
+}
