+++ lib/feature/data/i_datasource/i_remote_datasource/i_profile_remote_datasource.dart
@@ -1,0 +1,13 @@
+import 'package:cportal_flutter/feature/data/models/profile_model.dart';
+
+abstract class IProfileRemoteDataSource {
+  /// Обращается к эндпойнту .....
+  ///
+  /// Пробрасываем все ошибки через [ServerException]
+  Future<ProfileModel> getSingleProfile(String id, {bool isMyProfile = false});
+
+  /// Обращается к эндпойнту .....
+  ///
+  /// Пробрасываем все ошибки через [ServerException]
+  Future<List<ProfileModel>> searchProfiles(String query);
+}
