+++ lib/feature/data/i_datasource/i_remote_datasource/i_user_remote_datasource.dart
@@ -1,0 +1,8 @@
+import 'package:cportal_flutter/feature/data/models/user_model.dart';
+
+abstract class IUserRemoteDataSource {
+  /// Обращается к эндпойнту .....
+  ///
+  /// Пробрасываем все ошибки через [ServerException]
+  Future<UserModel> login(String connectongCode);
+}
