+++ lib/feature/data/repositories/news_repository_mobile.dart
@@ -1,8 +1,8 @@
 import 'package:cportal_flutter/core/error/cache_exception.dart';
 import 'package:cportal_flutter/core/error/server_exception.dart';
 import 'package:cportal_flutter/core/platform/i_network_info.dart';
-import 'package:cportal_flutter/feature/data/datasources/news_datasource/news_local_datasource.dart';
-import 'package:cportal_flutter/feature/data/datasources/news_datasource/news_remote_datasource.dart';
+import 'package:cportal_flutter/feature/data/i_datasource/i_local_datasource/i_news_local_datasource.dart';
+import 'package:cportal_flutter/feature/data/i_datasource/i_remote_datasource/i_news_remote_datasource.dart';
 import 'package:cportal_flutter/feature/data/models/news_model.dart';
 import 'package:cportal_flutter/core/error/failure.dart';
 import 'package:cportal_flutter/feature/domain/entities/news_entity.dart';
