+++ lib/feature/data/repositories/pin_code_repository.dart
@@ -1,13 +1,13 @@
 import 'package:cportal_flutter/core/error/server_exception.dart';
 
 import 'package:cportal_flutter/core/error/failure.dart';
-import 'package:cportal_flutter/feature/data/datasources/pin_code_local_datasource.dart';
+import 'package:cportal_flutter/feature/data/i_datasource/i_local_datasource/i_pin_code_local_datasource.dart';
 import 'package:cportal_flutter/feature/domain/repositories/i_pin_code_repository.dart';
 
 import 'package:dartz/dartz.dart';
 
 class PinCodeRepository implements IPinCodeRepository {
-  final IPinCodeDataSource localDataSource;
+  final IPinCodeLocalDataSource localDataSource;
 
   PinCodeRepository({
     required this.localDataSource,
