+++ lib/feature/data/repositories/user_repository_web.dart
@@ -1,8 +1,8 @@
 import 'package:cportal_flutter/core/error/cache_exception.dart';
 import 'package:cportal_flutter/core/error/server_exception.dart';
-import 'package:cportal_flutter/feature/data/datasources/user_datasource/user_local_datasource.dart';
-import 'package:cportal_flutter/feature/data/datasources/user_datasource/user_remote_datasource.dart';
 import 'package:cportal_flutter/core/error/failure.dart';
+import 'package:cportal_flutter/feature/data/i_datasource/i_local_datasource/i_user_local_datasource.dart';
+import 'package:cportal_flutter/feature/data/i_datasource/i_remote_datasource/i_user_remote_datasource.dart';
 import 'package:cportal_flutter/feature/data/models/user_model.dart';
 import 'package:cportal_flutter/feature/domain/repositories/i_user_repository.dart';
 import 'package:dartz/dartz.dart';
